// web-application/src/types-openapi/api/models/ListSliderRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListSliderRequest
 */
export interface ListSliderRequest {
    /**
     * 
     * @type {string}
     * @memberof ListSliderRequest
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListSliderRequest
     */
    status?: string | null;
}

/**
 * Check if a given object implements the ListSliderRequest interface.
 */
export function instanceOfListSliderRequest(value: object): value is ListSliderRequest {
    return true;
}

export function ListSliderRequestFromJSON(json: any): ListSliderRequest {
    return ListSliderRequestFromJSONTyped(json, false);
}

export function ListSliderRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListSliderRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'] == null ? undefined : json['title'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function ListSliderRequestToJSON(json: any): ListSliderRequest {
    return ListSliderRequestToJSONTyped(json, false);
}

export function ListSliderRequestToJSONTyped(value?: ListSliderRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'status': value['status'],
    };
}